requirements.txt
flask
pdfminer.six
python-docx
pandas
numpy
scikit-learn
transformers
torch
tqdm
nltk
joblib
sentence-transformers
spacy
pyresparser